fix(de1): reconcile bundled firmware with MMR exclusion
